Safari-609.1.11.1/LayoutTests/js/dfg-arrayify-when-prevent-extensions-expected.txt

Tests that Arraify does good things when Object.preventExtensions() has been called.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS var o = {}; Object.preventExtensions(o); foo(o) is void 0 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
